Cleanup of Exportable protocol.
