+++ WordPress/Classes/Extensions/ExportableAsset.swift
@@ -2,15 +2,8 @@ import Foundation
 
 @objc public protocol ExportableAsset: NSObjectProtocol {
 
-    func originalUTI() -> String?
-
     /// The MediaType for the asset
     ///
     var assetMediaType: MediaType { get }
-
-    /// The default UTI for thumbnails
-    ///
-    var defaultThumbnailUTI: String { get }
-
-    var mediaName: String? { get }
+    
 }
